style: Send view to view and discard statistical methods
