+++ ImageFeed/Screens/ImagesList/View/CellViewModel.swift
